Reorganize assistant-message
