+++ src/core/assistant-message/handlers/ToolDirectiveHandler.ts
@@ -1,6 +1,6 @@
 import * as sax from "sax"
 import { BaseDirectiveHandler } from "./BaseDirectiveHandler"
-import { ParseContext } from "../interfaces/ParseContext"
+import { ParseContext } from "../ParseContext"
 import { ToolUse, ToolParamName } from "../../../shared/tools"
 
 export class ToolDirectiveHandler extends BaseDirectiveHandler {
