--- src/core/assistant-message/index_original.ts
@@ -1,3 +1,0 @@
-export { type AssistantMessageContent, parseAssistantMessage } from "./parseAssistantMessage"
-export { presentAssistantMessage } from "./presentAssistantMessage"
-export { type LogDirective } from "./directives/LogDirective"
