--- src/core/assistant-message/parsers/index.ts
@@ -1,4 +1,0 @@
-export { TextContentParser } from "./TextContentParser"
-export { ToolUseParser } from "./ToolUseParser"
-export { ParameterParser } from "./ParameterParser"
-export type { ToolDirective, Directive, ParsingState } from "./types"
